equip order, equipment level ups
opacity updated to full when equipped, ordered equip by increments of 1 or same equipment slot, not by name of equipment enum. Equipment now levels up, I LEARNED SCRIPTABLE OBJECTS ONLY HAVE ONE INSTANCE, so that was hours of pain. For next time, remember when exactly you are modifying the values of a scriptable object and when you are accessing those values.