built, the fixing issues
Tried fixing rotation issues by changing it in code instead of on the inspector.

Added a check in LevelUp to try again if threeItems in the same.

Item slot is larger and has an itemName field now.

Updated targeting computer and body armor sprites.

Tried updating AudioManager to mute sounds on mainmenu so it doesn't play sounds from previous scene

Flamethrower has its own firepoint so it doesn't move the other weapons' firepoints