bosco updated, instance vs class issue
on level 1 bosco, adds a modifier of 1 to CharacterStats revives stat. Added Continue method to gameManager, will be called from the game over screen button "Continue?". UI checks for amount of revives, then removes one from character stats if used. Had an issue where the continue button was calling the class method instead of the GameManager.instance.continue method. Added the GameManagerReference script that find it and assigns it properly.

Also the button continue was not being assigned the method properly, moved the assigning a listener to the Start method instead of the awake and it worked.