level up issues
Changed the logic to calculate the level thresholds, now see that there are issues with the tracking of player xp vs the xp required for the next level up. For example, if the next level needs 25 xp to level up and the player has 15 xp, then it would start at 15 and not 0 - 25xp.